Deprecate min_segments parameter
