+++ wise_pizza/explain.py
@@ -27,7 +27,7 @@ def explain_changes_in_average(
     total_name: str,
     size_name: str,
     min_segments: Optional[int] = None,
-    max_segments: int = 5,
+    max_segments: int = None,
     min_depth: int = 1,
     max_depth: int = 2,
     solver: str = "lasso",
@@ -124,7 +124,7 @@ def explain_changes_in_totals(
     total_name: str,
     size_name: str,
     min_segments: Optional[int] = None,
-    max_segments: int = 5,
+    max_segments: int = None,
     min_depth: int = 1,
     max_depth: int = 2,
     solver: str = "lasso",
@@ -271,7 +271,7 @@ def explain_levels(
     total_name: str,
     size_name: Optional[str] = None,
     min_segments: int = None,
-    max_segments: int = 10,
+    max_segments: int = None,
     min_depth: int = 1,
     max_depth: int = 2,
     solver="lasso",
@@ -353,7 +353,7 @@ def explain_timeseries(
     time_name: str,
     size_name: Optional[str] = None,
     min_segments: int = None,
-    max_segments: int = 5,
+    max_segments: int = None,
     min_depth: int = 1,
     max_depth: int = 2,
     solver: str = "omp",
